Corrected and pre-tested version of power and update module
